Delete user's logged items when deleting user

+++ user_info/src/core/manage_user_info.py
@@ -1,3 +1,5 @@
+import os
+import requests
 from bson import ObjectId
 from user_info.src.models.converters.user_info_converter import UserInfoConverter
 from user_info.src.models.entities.user_info import UserInfo
@@ -23,6 +25,12 @@ def get_user_info_by_username(username: str) -> UserInfo | None:
         return None
 
 def delete_user(user: UserInfo):
+
+    # Delete all user's logged items
+    response: requests.Response = requests.delete(f"{os.environ['BACKEND_URL']}/logged_item/user/{str(user.pk)}")
+    if not response.ok:
+        raise Exception(f"Failed to delete user's logged items: {response.status_code=}, {response.text=}")
+    # Delete user
     return user.delete()
 
 def create_user(data: dict[str, str]) -> UserInfo:
